fixing auto-test errors.

+++ internal/gophermart/models/models.go
@@ -24,15 +24,15 @@ type Order struct {
 	Uploaded time.Time `json:"uploaded_at" db:"uploaded_at"`
 	Number   string    `json:"number" db:"number"`
 	Status   string    `json:"status" db:"status"`
-	Accrual  int64     `json:"accrual,omitempty" db:"accrual"`
+	Accrual  float32   `json:"accrual,omitempty" db:"accrual"`
 }
 
 type Users []User
 
 type User struct {
-	UserID   string `json:"login"`
-	Password string `json:"password"`
-	Accrual  int64  `json:"-"`
+	UserID   string  `json:"login"`
+	Password string  `json:"password"`
+	Accrual  float32 `json:"-"`
 }
 
 type Claims struct {
@@ -41,9 +41,9 @@ type Claims struct {
 }
 
 type AccrualResponse struct {
-	Status  string `json:"status"`
-	Number  string `json:"order"`
-	Accrual int64  `json:"accrual"`
+	Status  string  `json:"status"`
+	Number  string  `json:"order"`
+	Accrual float32 `json:"accrual"`
 }
 
 type Withdrawals []Withdrawal
@@ -52,10 +52,10 @@ type Withdrawal struct {
 	Processed time.Time `json:"processed_at" db:"processed_at"`
 	UserID    string    `json:"-" db:"userid"`
 	Number    string    `json:"order" db:"number"`
-	Sum       int64     `json:"sum" db:"sum"`
+	Sum       float32   `json:"sum" db:"sum"`
 }
 
 type Balance struct {
-	Current   int64 `json:"current"`
-	Withdrawn int64 `json:"withdrawn"`
+	Current   float32 `json:"current"`
+	Withdrawn float32 `json:"withdrawn"`
 }
